Trying make charts

+++ Backend/TEMPLATE_APP.WebApp/Dto/WeekStatistics.cs
@@ -8,18 +8,18 @@ namespace TEMPLATE_APP.WebApp.Dto
 
         public DateTime WeekDate { get; set; }
 
-        public int CommitsCount { get; set; }
+        public long CommitsCount { get; set; }
 
-        public int AdditionsLinesCount { get; set; }
+        public long AdditionsLinesCount { get; set; }
 
-        public int DeletionsLinesCount { get; set; }
+        public long DeletionsLinesCount { get; set; }
 
-        public int NewLinesCount { get; set; }
+        public long NewLinesCount { get; set; }
 
-        public int TotalLinesCount { get; set; }
+        public long TotalLinesCount { get; set; }
 
-        public int TotalCommitsCount { get; set; }
+        public long TotalCommitsCount { get; set; }
 
-        public int AuthorsCount { get; set; }
+        public long AuthorsCount { get; set; }
     }
 }
